# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login a user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Logout a user",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/admin/referees/{id}": {
      "get": {
        "operationId": "AdminRefereeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminReferee"
        ]
      },
      "patch": {
        "operationId": "AdminRefereeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRefereeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminReferee"
        ]
      },
      "delete": {
        "operationId": "AdminRefereeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminReferee"
        ]
      }
    },
...
    "/referees/{id}": {
      "get": {
        "operationId": "HomeRefereeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeReferee"
        ]
      }
    },
...
    "/admin/manager-types/{id}": {
      "get": {
        "operationId": "AdminManagerTypeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagerType"
        ]
      },
      "patch": {
        "operationId": "AdminManagerTypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SportManagerTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagerType"
        ]
      },
      "delete": {
        "operationId": "AdminManagerTypeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagerType"
        ]
      }
    },
...
    "/manager-types": {
      "get": {
        "operationId": "HomeManagerTypeController_findAll",
        "parameters": [
          {
            "name": "sportId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "active",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeManagerType"
        ]
      }
    },
    "/manager-types/sport/{sportId}": {
      "get": {
        "operationId": "HomeManagerTypeController_findAllBySportId",
        "parameters": [
          {
            "name": "sportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeManagerType"
        ]
      }
    },
    "/manager-types/{id}": {
      "get": {
        "operationId": "HomeManagerTypeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HomeManagerType"
        ]
      }
    },
...
    "/referee-types": {
      "get": {
        "operationId": "RefereeTypeHomeController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RefereeTypeHome"
        ]
      }
    },
    "/referee-types/{id}": {
      "get": {
        "operationId": "RefereeTypeHomeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RefereeTypeHome"
        ]
      }
    },
    "/referee-types/sport/{sportId}": {
      "get": {
        "operationId": "RefereeTypeHomeController_findAllBySportId",
        "parameters": [
          {
            "name": "sportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RefereeTypeHome"
        ]
      }
    },
    "/admin/referee-types": {
      "post": {
        "operationId": "RefereeTypeAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SportRefereeTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "RefereeTypeAdmin"
        ]
      }
    },
    "/admin/referee-types/{id}": {
      "patch": {
        "operationId": "RefereeTypeAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SportRefereeTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RefereeTypeAdmin"
        ]
      },
      "delete": {
        "operationId": "RefereeTypeAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RefereeTypeAdmin"
        ]
      }
    }
  },
  "info": {
    "title": "EZSportsRP API",
    "description": "API documentation",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "PASSword",
            "description": "The password of the user"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "The name of the user"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
...
      "MemberRefereeDto": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "number",
            "description": "Person ID"
          },
          "licenseNumber": {
            "type": "string",
            "description": "License number"
          },
          "licenseLevel": {
            "type": "string",
            "description": "License level"
          },
          "certification": {
            "type": "string",
            "description": "Certification"
          },
          "refereeType": {
            "type": "string",
            "description": "Referee type"
          },
          "specialization": {
            "type": "string",
            "description": "Specialization"
          },
          "languages": {
            "type": "string",
            "description": "Languages"
          },
          "matchFee": {
            "type": "number",
            "description": "Match fee"
          },
          "emergencyContactName": {
            "type": "string",
            "description": "Emergency contact name"
          },
          "emergencyContactPhone": {
            "type": "string",
            "description": "Emergency contact phone"
          },
          "status": {
            "type": "string",
            "description": "Status of the referee"
          },
          "bio": {
            "type": "string",
            "description": "Biography"
          },
          "ord": {
            "type": "number",
            "description": "Order of the referee"
          },
          "base64": {
            "type": "string",
            "description": "Base64 of the sport logo"
          }
        },
        "required": [
          "personId",
          "licenseNumber",
          "refereeType",
          "status",
          "ord"
        ]
      },
...
      "SportManagerTypeDto": {
        "type": "object",
        "properties": {
          "sportId": {
            "type": "number",
            "description": "Sport ID of the manager type"
          },
          "name": {
            "type": "string",
            "description": "Name of the manager type"
          },
          "abbr": {
            "type": "string",
            "description": "Abbreviation of the manager type"
          },
          "color": {
            "type": "string",
            "description": "Color of the manager type"
          },
          "note": {
            "type": "string",
            "description": "Note for the manager type"
          },
          "isActive": {
            "type": "number",
            "description": "Whether the manager type is active"
          },
          "ord": {
            "type": "number",
            "description": "Order of the manager type"
          },
          "base64": {
            "type": "string",
            "description": "Base64 of the sport logo"
          }
        },
        "required": [
          "name",
          "abbr",
          "color",
          "isActive",
          "ord"
        ]
      },
...
      "SportRefereeTypeDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}